requirements.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.2.0
cssutils==2.11.1
Django==5.1.6
django-anymail==12.0
djangorestframework==3.15.2
idna==3.10
jsonify==0.5
lxml==5.3.1
more-itertools==10.6.0
premailer==3.10.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
secure-smtplib==0.1.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
yagmail==0.15.293
